requirements.txt
chromadb
mdprint
fastapi==0.112.0 
fuzzywuzzy==0.18.0
langchain==0.2.14
langchain-chroma==0.1.2
langchain-openai==0.1.21
langchain-experimental==0.0.64 
langchain-community==0.2.12 
langchain-milvus==0.1.4
langchain-core==0.2.32
mdprint==0.2.0
pymilvus==2.4.5
pypdf==4.3.1
python-dotenv==1.0.1
redis==5.0.8
nest-asyncio==1.6.0
ipykernel